Update to convert the python int, which is long long, to size_t when needed

+++ cpp/csp/adapters/websocket/ClientOutputAdapter.cpp
@@ -6,7 +6,7 @@ ClientOutputAdapter::ClientOutputAdapter(
     Engine * engine,
     WebsocketEndpoint& endpoint,
     ClientAdapterManager * clientAdapterManager,
-    int64_t caller_id,
+    size_t caller_id,
     net::io_context& ioc
 ) : OutputAdapter( engine ), 
     m_endpoint( endpoint ),
